separate models into bagtiles and tiles
