+++ game/bagtile.py
@@ -1,10 +1,5 @@
 import random
-
-class Tile:
-    def __init__(self, letter, value):
-        self.letter = letter
-        self.value = value
-
+from game.tile import Tile
 
 class BagTiles:
     def __init__(self):
