+++ game/cell.py
@@ -1,10 +1,12 @@
-from game.models import Tile
+from game.tile import Tile
 
 class Cell:
-    def __init__(self, multiplier, multiplier_type = '', letter = None):
+    def __init__(self, multiplier:int , multiplier_type:bool, letter = None, active = bool):
         self.multiplier = multiplier
         self.multiplier_type = multiplier_type
         self.letter = letter
+        self.right_cell = None
+        self.active = active
 
     def add_letter(self, letter:Tile):
         self.letter = letter
@@ -12,7 +14,7 @@ class Cell:
     def calculate_value(self):
         if self.letter is None:
             return 0
-        if self.multiplier_type == 'letter':
+        if self.multiplier_type and self.active:
             return self.letter.value * self.multiplier
         else:
-            return self.letter.value        
+            return self.letter.value
