+++ game/scrabble.py
@@ -1,6 +1,7 @@
 from game.board import Board
 from game.player import Player
-from game.models import BagTiles
+from game.bagtile import BagTiles
+from game.tile import Tile
 
 class ScrabbleGame:
     def __init__(self, players_count):
