+++ game/tile.py
@@ -1,0 +1,6 @@
+class Tile:
+    def __init__(self, letter, value):
+        self.letter = letter
+        self.value = value
+
+
